Move Monotonic and Reboot zero-sized types into `board`
