+++ runners/lpc55/board/src/traits.rs
@@ -1,2 +1,6 @@
 pub mod buttons;
+pub mod monotonic;
+pub use monotonic::Monotonic;
+pub mod reboot;
+pub use reboot::Reboot;
 pub mod rgb_led;
