+++ runners/lpc55/board/src/traits/monotonic.rs
@@ -1,0 +1,1 @@
+pub use rtic::Monotonic;
