+++ runners/lpc55/board/src/traits/reboot.rs
@@ -1,0 +1,1 @@
+pub use admin_app::Reboot;
